+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.1 under development
 
-- no changes in this release.
+- Enh #100: Add support for multiple route hosts (Gerych1984)
 
 ## 1.0.0 December 30, 2021